requirements.txt
Flask==0.10
Markdown==2.5.2
google-cloud-bigquery==0.28.0
google-auth-oauthlib==0.2.0
oauth2client==4.1.2
BitstampClient==2.2.0
requests==2.3
Jinja2==2.10
